Read measurement even with health errors
Health errors are not always a blocker. They just tell that the
measurement may be inaccurate because some internal values within LIDAR
are not within "healthy" thresholds.

We still return the error, but allow the user, via type assertions, to
check that the error is health-related and thus may be ignored under
curtain curcumstances.